refactor: move /projects components to projects sections
